Remove votes alongside with chart
Fixed view for guests.

+++ app/views/charts/view.blade.php
@@ -13,7 +13,7 @@
         <td>{{$key+1}}</td>
         <td><a href="/charts/view/{{$chart->id}}">{{ucfirst($chart->type)}} Chart - {{$chart->name}}</a></td>
 		<td>
-			@if (Auth::user()->isAdmin())
+			@if (Auth::user()->isAdmin() && !Auth::guest())
 				<button type="button" onclick="window.location.href='/charts/delete/{{$chart->id}}'" class="btn btn-danger">Remove?</button> 
 			@endif
 		</td>
